+++ TheForest/SerializableTaskSystem/StoryCondition.cs
@@ -26,7 +26,7 @@ namespace TheForest.SerializableTaskSystem
 		{
 			if (!this._done)
 			{
-				GameStats.StoryElements storyElements = (GameStats.StoryElements)((int)o);
+				GameStats.StoryElements storyElements = (GameStats.StoryElements)o;
 				if (storyElements == this._type)
 				{
 					this.SetDone();
