+++ TheForest/TaskSystem/EnemyContactCondition.cs
@@ -28,7 +28,7 @@ namespace TheForest.TaskSystem
 			{
 				if (this._type != EnemyType.none)
 				{
-					if (this._type == (EnemyType)((int)o))
+					if (this._type == (EnemyType)o)
 					{
 						this.SetDone();
 						this.Clear();
